create face_creator.py and move methods around
